Fixed a bug with mpi communicating a float array as a double.
git-svn-id: https://pi-qmc.googlecode.com/svn/trunk@46 6df8c730-e442-11dd-8fb1-9960f2a117f8

+++ src/PairCFEstimator.h
@@ -97,7 +97,7 @@ public:
 #ifdef ENABLE_MPI
     if (mpi) {
       mpi->getWorkerComm().Reduce(temp.data(),mpiBuffer.data(),
-                                  product(nbin),MPI::DOUBLE,MPI::SUM,0);
+                                  product(nbin),MPI::FLOAT,MPI::SUM,0);
       temp = mpiBuffer;
     }
 #endif
